# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Anatomy Pro API
  version: 1.0.0
  description: API REST para o jogo Anatomy Pro
servers:
  - url: https://api.anatomypro.local
paths:
...
  /v1/users/me:
    get:
      summary: Obtém dados do usuário logado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      summary: Atualiza dados do usuário
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /v1/users/me/preferences:
    post:
      summary: Atualiza preferências do usuário
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceUpdate'
      responses:
        '200':
          description: Preferências atualizadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /v1/users/me/summary:
    get:
      summary: Resumo de progresso do usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resumo do perfil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSummary'
...
  /v1/quizzes/sessions:
    post:
      summary: Cria sessão de quiz
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizSessionCreate'
      responses:
        '201':
          description: Sessão iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSession'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/quizzes/sessions/{sessionId}/attempts:
    post:
      summary: Registra tentativa de pergunta
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizAttemptCreate'
      responses:
        '200':
          description: Tentativa registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizAttempt'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/quizzes/sessions/{sessionId}/complete:
    post:
      summary: Finaliza sessão de quiz
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                duration_seconds:
                  type: integer
                  minimum: 0
              required:
                - duration_seconds
      responses:
        '200':
          description: Sessão concluída
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSession'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        status_code:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        profile_type:
          type: string
        role:
          type: string
        xp:
          type: integer
        streak:
          type: integer
        energy:
          type: integer
        elo_rating:
          type: integer
        preferences:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserUpdate:
      type: object
      properties:
        display_name:
          type: string
        preferences:
          type: object
        energy:
          type: integer
          minimum: 0
    PreferenceUpdate:
      type: object
      properties:
        dark_mode:
          type: boolean
        language:
          type: string
        notifications:
          type: boolean
...
    ProfileSummary:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        systems_progress:
          type: object
          additionalProperties:
            type: number
        daily_missions_completed:
          type: integer
        weekly_missions_completed:
          type: integer
...
    QuizSessionCreate:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [sprint, campaign, osce, srs]
        system:
          type: string
          nullable: true
        difficulty:
          type: string
          nullable: true
        limit:
          type: integer
          default: 10
    QuizSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mode:
          type: string
        system:
          type: string
          nullable: true
        score:
          type: number
        duration_seconds:
          type: integer
        completed:
          type: boolean
    QuizAttemptCreate:
      type: object
      required: [question_id, option_id]
      properties:
        question_id:
          type: string
          format: uuid
        option_id:
          type: string
          format: uuid
    QuizAttempt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        selected_option_id:
          type: string
          format: uuid
          nullable: true
        is_correct:
          type: boolean
...